requirements.txt
streamlit
pandas
numpy
tabula-py
camelot-py
matplotlib
PyPDF2